requirements.txt
geopandas
rasterio
rasterstats
# numpy
# shapely
# fiona
# rasterio
# pyproj
# pytest >= 4.6
# rasterstats
